move lifetime to own file
