+++ ts/quickjs.test.ts
@@ -2,30 +2,10 @@
  * These tests demonstate some common patterns for using quickjs-emscripten.
  */
 
-import { getQuickJS, QuickJSVm, QuickJSHandle, InterruptHandler, Lifetime } from './quickjs'
+import { getQuickJS, QuickJSVm, QuickJSHandle, InterruptHandler } from './quickjs'
 import { it, describe } from 'mocha'
 import assert from 'assert'
 import { VmCallResult } from './vm-interface'
-
-describe('Lifetime', () => {
-  describe('.consume', () => {
-    it('yeilds the value', () => {
-      const lifetime = new Lifetime(1)
-      const result = lifetime.consume(l => l.value + 1)
-      assert.strictEqual(result, 2)
-    })
-
-    it('disposes the lifetime', () => {
-      let disposed = false
-      const lifetime = new Lifetime(2, undefined, () => {
-        disposed = true
-      })
-      lifetime.consume(l => l.value * 2)
-      assert.strictEqual(lifetime.alive, false)
-      assert.strictEqual(disposed, true)
-    })
-  })
-})
 
 describe('QuickJSVm', async () => {
   let vm: QuickJSVm = undefined as any
